pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "weavergen"
version = "0.1.0"
description = "Python wrapper for OTel Weaver Forge with Claude Code optimization"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Sean Chatman", email = "sean@seanchatman.com" },
]
keywords = [
    "opentelemetry",
    "semantic-conventions", 
    "code-generation",
    "observability",
    "weaver",
    "claude-code"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.1.0",
    "jinja2>=3.1.0", 
    "pyyaml>=6.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.0.0",
]

llm = [
    "ollama>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]

examples = [
    "asyncpg>=0.29.0",
]

otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
]

all = [
    "weavergen[dev,llm,otel,examples]",
    "icontract>=2.6.0",
]

[project.urls]
Homepage = "https://github.com/seanchatmangpt/weavergen"
Documentation = "https://github.com/seanchatmangpt/weavergen#readme"
Repository = "https://github.com/seanchatmangpt/weavergen.git"
"Bug Tracker" = "https://github.com/seanchatmangpt/weavergen/issues"

[project.scripts]
weavergen = "weavergen.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/weavergen"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
]

[tool.coverage.run]
source = ["src/weavergen"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
